feat: Add Writer Field in SearchBookmarkDTO, Refactor Directory [#75]
