+++ src/main/java/com/example/teampandanback/dto/note/response/NoteEachSearchInBookmarkResponse.java
@@ -1,4 +1,4 @@
-package com.example.teampandanback.dto.note.search;
+package com.example.teampandanback.dto.note.response;
 
 import com.example.teampandanback.domain.note.Step;
 import lombok.Getter;
@@ -10,12 +10,14 @@ public class NoteEachSearchInBookmarkResponse {
     private Step step;
     private Long projectId;
     private String projectTitle;
+    private String writer;
 
-    public NoteEachSearchInBookmarkResponse(Long noteId, String title, Step step, Long projectId, String projectTitle) {
+    public NoteEachSearchInBookmarkResponse(Long noteId, String title, Step step, Long projectId, String projectTitle, String writer) {
         this.noteId = noteId;
         this.title = title;
         this.step = step;
         this.projectId = projectId;
         this.projectTitle = projectTitle;
+        this.writer = writer;
     }
 }
