+++ src/main/java/com/example/teampandanback/service/BookmarkService.java
@@ -10,8 +10,8 @@ import com.example.teampandanback.domain.user.UserRepository;
 import com.example.teampandanback.domain.user_project_mapping.UserProjectMapping;
 import com.example.teampandanback.domain.user_project_mapping.UserProjectMappingRepository;
 import com.example.teampandanback.dto.auth.SessionUser;
-import com.example.teampandanback.dto.note.search.NoteEachSearchInBookmarkResponse;
-import com.example.teampandanback.dto.note.search.NoteSearchInBookmarkResponse;
+import com.example.teampandanback.dto.note.response.NoteEachSearchInBookmarkResponse;
+import com.example.teampandanback.dto.note.response.NoteSearchInBookmarkResponse;
 import com.example.teampandanback.exception.ApiRequestException;
 import com.example.teampandanback.utils.PandanUtils;
 import lombok.RequiredArgsConstructor;
